+++ s/layout/types.ts
@@ -3,6 +3,7 @@ export type Id = string
 
 export type LayoutOptions = {
 	stock: Stock
+	defaultPanel?: string
 }
 
 export type Blueprint = {
@@ -65,4 +66,3 @@ export type Stock = {
 	default: () => Cell
 	empty: () => Cell
 }
-