+++ s/studio/types.ts
@@ -2,9 +2,9 @@
 import {Kv} from "@e280/kv"
 import {Content} from "@e280/sly"
 
-import type {Studio} from "./studio.js"
 import {Layout} from "../layout/layout.js"
 import {Id, Surface, Dock} from "../layout/types.js"
+import {LayoutMeta} from "./ui/desk/rendering/utils/layout-meta.js"
 
 export type Renderer = (desk: HTMLElement) => (surfaces: Surface[]) => void
 
@@ -16,24 +16,24 @@ export type StandardButtonKey =
 	| "resetLayout"
 
 export type DockContext = {
-	studio: Studio<any>
 	dock: Dock
+	meta: LayoutMeta
 }
 
-export type DockButtonsFn = (context: DockContext) => Content
-export type StandardButtons = Record<StandardButtonKey, () => Content>
+export type DockControlsFn = (context: DockContext) => Content
 
 export type StudioOptions<Ps extends Panels> = {
 	layout: Layout
 	panels: Ps
 	renderer: Renderer
-	buttons?: DockButtonsFn
+	controls?: DockControlsFn
 }
 
 export type PersistenceOptions = {
 	layout: Layout
 	kv: Kv
 	key: string
+	broadcastChannel: BroadcastChannel
 }
 
 export type Panel = {