+++ s/studio/ui/desk/rendering/utils/layout-meta.ts
@@ -4,10 +4,12 @@ import {Studio} from "../../../../studio.js"
 import {Resizer} from "../../resize/resizer.js"
 import {LayoutNode} from "../../../../../layout/types.js"
 import {TabDragger} from "../../parts/tab-dragger.js"
+import {TaskbarDragger} from "../../parts/taskbar-dragger.js"
 
 export type LayoutMeta = {
 	resizer: Resizer
-	dragger: TabDragger
+	tabDragger: TabDragger
+	taskbarDragger: TaskbarDragger
 	studio: Studio
 	renderLayout: (node: LayoutNode) => TemplateResult
 }