+++ s/studio/ui/desk/rendering/utils/list-panels.ts
@@ -1,12 +1,9 @@
 
-import {html} from "lit"
 import {LayoutMeta} from "./layout-meta.js"
 import {Dock} from "../../../../../layout/types.js"
 
-export function renderAdderSurface(
-		{studio: {layout, panels, ...stuff}}: LayoutMeta,
-		dock: Dock,
-	) {
+export function listPanelsChoices(meta: LayoutMeta, dock: Dock) {
+	const {studio: {layout, panels}} = meta
 
 	const atLimit = (panel: string, limit = Infinity) => {
 		if (!Number.isFinite(limit))
@@ -19,21 +16,18 @@ export function renderAdderSurface(
 		return (panelCount >= limit)
 	}
 
-	function click(name: string) {
+	function open(name: string) {
 		return async() => {
 			const {index} = await layout.actions.addSurface(dock.id, name)
 			await layout.actions.setDockActiveSurface(dock.id, index)
 		}
 	}
 
-	return Object.entries(panels).map(([name, panel]) => html`
-		<button
-			@click="${click(name)}"
-			?disabled="${atLimit(name, panel.limit)}">
-
-			${panel.icon({studio: {layout, panels, ...stuff}, dock})}
-			<span>${panel.label}</span>
-		</button>
-	`)
+	return Object.entries(panels).map(([name, panel]) => ({
+		name,
+		label: panel.label,
+		icon: panel.icon({meta, dock}),
+		disabled: atLimit(name, panel.limit),
+		open: open(name),
+	}))
 }
-